[feat] #123 ChatDetail 및 Chat에 toUiModel() 작성

+++ data/src/main/java/com/mashup/data/repository/ChatRepositoryImpl.kt
@@ -9,7 +9,7 @@ class ChatRepositoryImpl @Inject constructor(
     private val remoteChatDataSource: RemoteChatDataSource
 ): ChatRepository {
 
-    override suspend fun getChatRooms(id: Int): ChatDetail {
+    override suspend fun getChatRooms(id: Long): ChatDetail {
         return remoteChatDataSource.getChats(id).toDomainModel()
     }
 }
