+++ data/src/main/java/com/mashup/data/source/remote/service/ChatService.kt
@@ -9,6 +9,6 @@ interface ChatService {
 
     @GET("rooms/{id}/chats")
     suspend fun getChats(
-        @Path("id") id: Int
+        @Path("id") id: Long
     ): BaseResponse<GetChatsResponseBody>
 }
